add support for the template generating a top-level array of items
When registering a template to mock, you may optionally pass in a boolean is_array parameter.
If set, mockJSON will assume the template is an object with a single child property.
The key is a name, which is discarded, except for the range, which specifies the number of elements to generate.
The value is an array template.